Move URLMap to utils
