+++ pkg/utils/urlmap.go
@@ -1,4 +1,4 @@
-package config
+package utils
 
 import "strings"
